requirements.txt
aws-cdk-lib>=2.135.0
constructs>=10.3.0
requests>=2.31.0
